% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{check_submission}
\alias{check_submission}
\title{Check that submission was successfully processed}
\usage{
check_submission(forecast_file)
}
\arguments{
\item{forecast_file}{Your forecast csv or nc file}
}
\description{
Check that submission was successfully processed
}
