% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{combined_scores}
\alias{combined_scores}
\title{Combining scores from a theme together}
\usage{
combined_scores(theme, collect = TRUE)
}
\arguments{
\item{theme}{theme name}

\item{collect}{TRUE/FALSE to download results}
}
\value{
a data.frame of scores
}
\description{
Combining scores from a theme together
}
