% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{create_model_metadata}
\alias{create_model_metadata}
\title{Write model metadata from template}
\usage{
create_model_metadata(forecast_file)
}
\arguments{
\item{forecast_file}{full path of forecast file}
}
\description{
Write model metadata from template
}
\examples{
    create_model_metadata(forecast_file = forecast_file)                          
}
