% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_forecast.R
\name{download_forecast}
\alias{download_forecast}
\title{Download forecasts for NEON sites from the EFI server}
\usage{
download_forecast(
  theme,
  date = Sys.Date() - 2,
  dir = tempdir(),
  s3_region = "data",
  s3_endpoint = "ecoforecast.org"
)
}
\arguments{
\item{theme}{string of the theme}

\item{date}{start date for the forecast}

\item{dir}{storage location.  Use tempdir unless you want to keep this
data around on your computer, in which case, \code{neonstore::neon_dir()} might
be a convenient choice.}

\item{s3_region}{data}

\item{s3_endpoint}{ecoforecast.org}
}
\description{
Download forecasts for NEON sites from the EFI server
}
\examples{
download_forecast("phenology")
}
