% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{download_noaa}
\alias{download_noaa}
\title{Download NOAA Weather forecasts for NEON sites from the EFI server}
\usage{
download_noaa(
  siteID,
  interval = "6hr",
  date = Sys.Date() - 2,
  cycle = "00",
  dir = tempdir()
)
}
\arguments{
\item{siteID}{vector of 4-character NEON siteIDs}

\item{interval}{Time interval for the forecast}

\item{date}{start date for the forecast}

\item{cycle}{NOAA hour-cycle (first hour of the forecast),
options are "00", "06", "12", "18"; only the "00" forecasts
run 35 days into future.}

\item{dir}{storage location.  Use tempdir unless you want to keep this
data around on your computer, in which case, \code{neonstore::neon_dir()} might
be a convenient choice.}
}
\description{
Download NOAA Weather forecasts for NEON sites from the EFI server
}
