% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_helpers.R
\name{efi_format}
\alias{efi_format}
\title{Format a fable fbl_ts forecast in EFI standard}
\usage{
efi_format(df, times = 10)
}
\arguments{
\item{df}{a fbl_ts forecast}

\item{times}{number of times to draw from distribution for ensemble method,
ignored if distributions are normal.}
}
\value{
A data.frame (\verb{[tsibble]}) in the EFI format
}
\description{
Format a fable fbl_ts forecast in EFI standard
}
\examples{

\dontrun{
library(dplyr)
library(readr)
library(fable)
aquatic <-
  read_csv("https://data.ecoforecast.org/neon4cast-targets/aquatics/aquatics-targets.csv.gz") \%>\%
  pivot_wider(names_from = "variable", values_from = "observation") \%>\%
  as_tsibble(index = datetime, key = site_id)
oxygen_fc <- aquatic \%>\%
  model(null = MEAN(oxygen)) \%>\%
  forecast(h = "35 days") \%>\%
  efi_format()
 }
}
