% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_helpers.R
\name{efi_format_ensemble}
\alias{efi_format_ensemble}
\title{Format as EFI using ensemble draws}
\usage{
efi_format_ensemble(df, times = 10)
}
\arguments{
\item{df}{a fbl_ts forecast}

\item{times}{number of ensemble members to draw
Supports unrecognized distributions by drawing samples}
}
\description{
Format as EFI using ensemble draws
}
