% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scores.R
\name{fill_scores}
\alias{fill_scores}
\title{fill_scores}
\usage{
fill_scores(df, null_team = "EFInull")
}
\arguments{
\item{df}{data.frame of raw scores.  Must have columns:
\code{theme}, \code{siteID}, \code{target}, \code{time}, \code{team}, \code{forecast_start_time},
\code{crps} & \code{logs}.
@param null name of the null team}
}
\description{
To compare teams which issue forecasts of multiple observations, we need
a rule for dealing with cases where one team makes predictions for
certain observations at a given forecast horizon when another team does not.
For instance, teams may use different horizon lengths (forecast different
number of days into the future).  Alternately, two teams may both predict
the same observation (target, time, siteID) but the one team does so
much farther in advance than the other.  These miss-matches create
"implicit" missing observations relative to other scores.  If we simply omit
missing values, teams have little incentive to make a full set of predictions
-- dropping missing values while making one good prediction could lead to a much
much better score.
}
\details{
To address this, we use a null score as a reference.  Teams are scored for all observations
and forecast horizons for which the null forecast contains a prediction.
If the team has made a forecast of a given observation but only on a longer horizon,
we will re-use that prediction when comparing against the prediction on the shorter horizon as well.
If no prior forecast is available, we will fill the missing prediction with that of the null
forecast.
}
