% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_output_validator.R
\name{forecast_output_validator}
\alias{forecast_output_validator}
\title{Validate forecast file}
\usage{
forecast_output_validator(forecast_file)
}
\arguments{
\item{forecast_file}{forecast csv or csv.gz file}
}
\description{
Validate forecast file
}
