% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{generate_metadata}
\alias{generate_metadata}
\title{Generate metadata from forecast and template files}
\usage{
generate_metadata(
  forecast_file,
  metadata_yaml,
  forecast_issue_time,
  forecast_iteration_id
)
}
\arguments{
\item{forecast_file}{full path to forecast file}

\item{metadata_yaml}{full path to meta data template fill}

\item{forecast_issue_time}{time that forecast was generated}

\item{forecast_iteration_id}{unique ID for forecast}
}
\description{
Generate metadata from forecast and template files
}
\examples{
use `EML` package to build the attribute list
use `EML` package to build the physical list
use `EML` package to dataTable
Create the coverage EML
Create the dataset EML
Check base EML
Check that EML matches EFI Standards
Write metadata
}
