% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{get_noaa_forecast_s3}
\alias{get_noaa_forecast_s3}
\title{Download driver forecasts from s3 bucket}
\usage{
get_noaa_forecast_s3(
  dir,
  model,
  site,
  date,
  cycle,
  s3_region = Sys.getenv("AWS_DEFAULT_REGION")
)
}
\arguments{
\item{dir}{full path to working directory}

\item{forecast_path}{relative path of the driver forecast (relative to driver directory or bucket)}
}
\value{

}
\description{
Download driver forecasts from s3 bucket
}
