% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{get_stacked_noaa_s3}
\alias{get_stacked_noaa_s3}
\title{Download stacked NOAA data from s3 bucket}
\usage{
get_stacked_noaa_s3(
  dir,
  site,
  averaged = TRUE,
  s3_region = Sys.getenv("AWS_DEFAULT_REGION")
)
}
\arguments{
\item{dir}{full path to working directory}

\item{averaged}{logistical; TRUE = download averaged stacked forecast}

\item{config}{flare configuration object}
}
\value{

}
\description{
Download stacked NOAA data from s3 bucket
}
