% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_helpers.R
\name{get_target}
\alias{get_target}
\title{Download target data from s3}
\usage{
get_target(dir, theme, s3_region = Sys.getenv("AWS_DEFAULT_REGION"))
}
\arguments{
\item{dir}{full path to working directory}

\item{theme}{forecast theme}

\item{s3_region}{s3 region}
}
\description{
Download target data from s3
}
