% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scores.R
\name{mean_scores}
\alias{mean_scores}
\title{mean_scores}
\usage{
mean_scores(df)
}
\arguments{
\item{df}{a data frame from fill_scores()}
}
\description{
Compute the mean scores from \code{\link[=fill_scores]{fill_scores()}} table.
By default, results are sorted by mean_crps,
based on self-fill and then null-fill.  For comparisons
across different horizons, scores on null_filled_crps alone
(no self-fill step) may be preferred.
The number of missing values filled in for each forecast is also reported.
}
