% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs.R
\name{noaa_stage3}
\alias{noaa_stage3}
\title{NOAA GEFS forecasts with EFI stage 3 processing}
\usage{
noaa_stage3(
  version = "v12",
  endpoint = "data.ecoforecast.org",
  verbose = TRUE,
  site_id = NA
)
}
\arguments{
\item{version}{GEFS forecast version. Prior versions correspond to forecasts
issued before 2020-09-25 which have different ensemble number and horizon,
among other changes, and are not made available here. Leave as default.}

\item{endpoint}{the EFI host address (leave as default)}

\item{verbose}{logical, displays or hides messages}

\item{cycle}{Hour at which forecast was made, as character string
(\code{"00"}, \code{"06"}, \code{"12"} or \code{"18"}). Only \code{"00"} (default) has 30 days horizon.}
}
\description{
Stage 3 processing presents a 'nowcast' product by combining the most
recent predictions from each available cycle. Product uses CF variable
names and 1 hr interval
}
