% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score}
\alias{score}
\title{Compute the CRPS score of your forecast}
\usage{
score(
  forecast,
  theme = c("aquatics", "beetles", "phenology", "terrestrial_30min",
    "terrestrial_daily", "ticks")
)
}
\arguments{
\item{forecast}{forecast data frame or file}

\item{theme}{theme name. Note: terrestrial must specify the interval.}
}
\description{
Compute the CRPS score of your forecast
}
