% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_schema}
\alias{score_schema}
\title{score schema for arrow::open_dataset}
\usage{
score_schema()
}
\description{
Open all score.csv files from local or remote disk using arrow by
declaring the appropriate file schema.
}
