% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{submit}
\alias{submit}
\title{submit forecast to EFI}
\usage{
submit(
  forecast_file,
  metadata = NULL,
  ask = interactive(),
  s3_region = "data",
  s3_endpoint = "ecoforecast.org"
)
}
\arguments{
\item{forecast_file}{Your forecast csv or nc file}

\item{metadata}{path to metadata file}

\item{ask}{should we prompt for a go before submission?}

\item{s3_region}{subdomain (leave as is for EFI challenge)}

\item{s3_endpoint}{root domain (leave as is for EFI challenge)}
}
\description{
submit forecast to EFI
}
