% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{theme_statistics}
\alias{theme_statistics}
\title{Calculating forecast challenge submission statistics}
\usage{
theme_statistics(themes)
}
\arguments{
\item{themes}{theme names}
}
\value{
a data.frame of challenge statistics
}
\description{
Calculating forecast challenge submission statistics
}
