% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_forecast}
\alias{validate_forecast}
\title{validate a forecast CSV or NetCDF file, or print helpful error messages}
\usage{
validate_forecast(forecast_file)
}
\description{
validate a forecast CSV or NetCDF file, or print helpful error messages
}
