% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_meta_template}
\alias{write_meta_template}
\title{Write metadata from template}
\usage{
write_meta_template(forecast_file)
}
\arguments{
\item{forecast_file}{full path of forecast file}
}
\description{
Write metadata from template
}
