% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_metadata_eml}
\alias{write_metadata_eml}
\title{Generate metadata from forecast and template files}
\usage{
write_metadata_eml(
  forecast_file,
  metadata_yaml,
  forecast_issue_time,
  forecast_iteration_id
)
}
\arguments{
\item{forecast_file}{full path to forecast file}

\item{metadata_yaml}{full path to meta data template fill}

\item{forecast_issue_time}{time that forecast was generated}

\item{forecast_iteration_id}{unique ID for forecast}
}
\description{
Generate metadata from forecast and template files
}
