% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{check_submission}
\alias{check_submission}
\title{Check that submission was successfully processed}
\usage{
check_submission(
  forecast_file,
  s3_region = "data",
  s3_endpoint = "ecoforecast.org"
)
}
\arguments{
\item{forecast_file}{Your forecast csv or nc file}

\item{s3_region}{subdomain (leave as is for EFI challenge)}

\item{s3_endpoint}{root domain (leave as is for EFI challenge)}
}
\description{
Check that submission was successfully processed
}
