% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_helpers.R
\name{efi_format}
\alias{efi_format}
\title{Format a fable fbl_ts forecast in EFI standard}
\usage{
efi_format(df, times = 100)
}
\arguments{
\item{df}{a fbl_ts forecast}

\item{times}{number of times to draw from distribution for ensemble method,
ignored if distributions are normal.}
}
\value{
A data.frame (\verb{[tsibble]}) in the EFI format
}
\description{
Format a fable fbl_ts forecast in EFI standard
}
\examples{
library(dplyr)
library(readr)
library(fable)
aquatic <- read_csv("https://data.ecoforecast.org/targets/aquatics/aquatics-targets.csv.gz") \%>\% 
as_tsibble(index=time, key=siteID)
oxygen_fc <- aquatic \%>\%
  model(null = MEAN(oxygen)) \%>\%
  forecast(h = "35 days") \%>\%
  efi_format()

}
