% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_output_validator.R
\name{forecast_output_validator}
\alias{forecast_output_validator}
\title{forecast_output_validator}
\usage{
forecast_output_validator(
  file_in,
  grouping_variables = c("siteID", "time"),
  target_variables = c("oxygen", "temperature", "richness", "abundance", "nee", "le",
    "vswc", "gcc_90"),
  theme_names = c("aquatics", "beetles", "phenology", "terrestrial_30min",
    "terrestrial_daily", "ticks")
)
}
\arguments{
\item{file_in}{}

\item{grouping_variables}{}

\item{target_variables}{}
}
\description{
forecast_output_validator
}
