% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_output_validator.R
\name{forecast_output_validator}
\alias{forecast_output_validator}
\title{forecast_output_validator}
\usage{
forecast_output_validator(
  forecast_file,
  grouping_variables = c("siteID", "time"),
  target_variables = c("oxygen", "temperature", "richness", "abundance", "nee", "le",
    "vswc", "gcc_90", "rcc_90", "ixodes_scapularis", "amblyomma_americanum"),
  theme_names = c("aquatics", "beetles", "phenology", "terrestrial_30min",
    "terrestrial_daily", "ticks")
)
}
\arguments{
\item{forecast_file}{Your forecast csv or nc file}

\item{grouping_variables}{Grouping variables}

\item{target_variables}{Possible target variables}

\item{theme_names}{valid EFI theme names}
}
\description{
forecast_output_validator
}
\examples{

forecast_file <- system.file("extdata/aquatics-2021-02-01-EFInull.csv.gz", 
                              package = "neon4cast")
forecast_output_validator(forecast_file)

}
