% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{generate_metadata}
\alias{generate_metadata}
\title{Generate metadata from forecast and template files}
\usage{
generate_metadata(
  forecast_file,
  team_list,
  model_metadata,
  forecast_iteration_id = NULL,
  title = NULL,
  team_name = NULL,
  intellectualRights = "https://creativecommons.org/licenses/by/4.0/",
  abstract = NULL,
  methods = NULL,
  attributes = NULL,
  forecast_issue_time = NULL
)
}
\arguments{
\item{forecast_file}{full path to forecast file}

\item{model_metadata}{list of model metadata}

\item{forecast_iteration_id}{unique ID for forecast}

\item{title}{,}

\item{list}{of team members}
}
\description{
Generate metadata from forecast and template files
}
