% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs.R
\name{noaa_stage2}
\alias{noaa_stage2}
\title{NOAA GEFS forecasts with EFI stage 2 processing
Stage2 processing involves the following transforms of the data:
\itemize{
\item Fluxes are standardized to per second rates
\item Variables are renamed to match CF conventions
\item Fluxes and states are interpolated to 1 hour intervals
}}
\usage{
noaa_stage2(
  cycle = 0,
  version = "v12",
  endpoint = "data.ecoforecast.org",
  verbose = TRUE
)
}
\arguments{
\item{cycle}{Hour at which forecast was made, as character string
(\code{"00"}, \code{"06"}, \code{"12"} or \code{"18"}). Only \code{"00"} (default) has 30 days horizon.}

\item{version}{GEFS forecast version. Prior versions correspond to forecasts
issued before 2020-09-25 which have different ensemble number and horizon,
among other changes, and are not made available here. Leave as default.}

\item{endpoint}{the EFI host address (leave as default)}

\item{verbose}{logical, displays or hides messages}
}
\description{
NOAA GEFS forecasts with EFI stage 2 processing
Stage2 processing involves the following transforms of the data:
\itemize{
\item Fluxes are standardized to per second rates
\item Variables are renamed to match CF conventions
\item Fluxes and states are interpolated to 1 hour intervals
}
}
