% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score}
\alias{score}
\title{Compute the CRPS score of your forecast}
\usage{
score(
  forecast,
  theme = c("aquatics", "beetles", "phenology", "terrestrial_30min",
    "terrestrial_daily", "ticks"),
  target_url = NA,
  target_vars = c("oxygen", "temperature", "richness", "abundance", "nee", "le",
    "vswc", "gcc_90", "rcc_90", "ixodes_scapularis", "amblyomma_americanum")
)
}
\arguments{
\item{forecast}{forecast data frame or file}

\item{theme}{theme name. Note: terrestrial must specify the interval.}
}
\description{
Compute the CRPS score of your forecast
}
\examples{
forecast_file <- system.file("extdata/aquatics-2021-02-01-EFInull.csv.gz", 
                              package = "neon4cast")
score(forecast_file, "aquatics")                          
}
