% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{submit}
\alias{submit}
\title{submit forecast to EFI}
\usage{
submit(forecast_file, metadata = NULL, ask = interactive())
}
\arguments{
\item{forecast_file}{Your forecast csv or nc file}

\item{metadata}{path to metadata file}

\item{ask}{should we prompt for a go before submission?}
}
\description{
submit forecast to EFI
}
